# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Website Profile Disbudporapar API",
        "description": "API untuk Elektronik Survei Kepuasan Masyarakat (E-SKM) Pemerintah Kabupaten Katingan. Digunakan untuk mengelola autentikasi pengguna, data pengguna, dan survei kepuasan terhadap layanan perangkat daerah.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/organizations": {
            "get": {
                "tags": [
                    "Organization"
                ],
                "summary": "Retrieve a list of Organizations",
                "operationId": "38539e67fb3c20065e66e337a5fc6a37",
                "responses": {
                    "200": {
                        "description": "Organization retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Organization"
                ],
                "summary": "Create a new Organization",
                "description": "Create a new Organization",
                "operationId": "705373d2182b3af62bdb4bed5c21fb1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Organizaion created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Organizaion created successfully"
                                        },
                                        "organization": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "name_value"
                                                },
                                                "position": {
                                                    "type": "string",
                                                    "example": "position_value"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token invalid or missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "key": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The key field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create Organizaion"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Database error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}